+++ R/utils.R
@@ -62,7 +62,7 @@ ept <- function(x, envir = parent.frame()) {
 # Removes object's attributes before printing
 print_no_attr <- function(x, ...) {
   attributes(x) <- NULL
-  NextMethod("print", x)
+  print(x)
 }
 
 #' That