+++ NEWS.md
@@ -3,6 +3,11 @@
 * `match_param()` now reports the value passed to `param` on failure
 * removes checks for `stringsAsFactors` option in `quick_df()`
 * adds `print_c()` to print a vector into a `c()` "string"
+* improvements to handlers
+  * adds `has_message()` and `get_message()`
+  * internal rework of `catch()` for catching errors, messages, and warnings
+  * `has_*()` now returns the result in the `result` attribute
+  * `print.has_catch()` will hide attributes in print
 
 # jordan 0.1.1